pyproject.toml
[project]
name = "watchit"
version = "0.1.0"
description = "Local parental monitoring with agentic LLM judge (Ollama)"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.111.0,<1",
  "hypercorn>=0.16.0",
  "uvloop>=0.19.0; platform_system!='Windows'",
  "pydantic>=2.6,<3",
  "pydantic-settings>=2.4,<3",
  "sqlalchemy>=2.0,<3",
  "aiosqlite>=0.20",
  "orjson>=3.10",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "sse-starlette>=2.1.0",
  "langchain>=0.2.7,<0.3",
  "langchain-community>=0.2.7,<0.3",
  "langgraph>=0.2.30,<0.3",
  "pysqlcipher3-binary>=1.0.4",
  "pytesseract>=0.3.10",
  "faster-whisper>=1.0",
  "Pillow>=10.3", 
  # CORS for dashboard
  "starlette>=0.37",
]

[tool.ruff]
line-length = 100
target-version = "py310"

